+++ config/menu_seeder/masterdata/masterdata-usermanagement_seeder.php
@@ -50,5 +50,13 @@ return [
             'permissions' => ['create', 'read', 'update', 'delete'],
             'roles' => ['admin', 'master'],
         ],
+        [
+            'title' => 'Data Login',
+            'title_en' => 'Login Data',
+            'title_key' => 'md_data_login',
+            'route' => 'usermanagement.data-login',
+            'permissions' => ['read'],
+            'roles' => ['admin', 'master'],
+        ],
     ],
 ];